_All_Projects/12.11/requirements.txt
Django>=3,<4
djangorestframework

